apps/api/pyproject.toml
[tool.poetry]
name = "gstgpt-api"
version = "0.1.0"
description = "GST GPT Backend API"
authors = ["Admin <admin@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
sqlalchemy = "^2.0.25"
alembic = "^1.13.1"
asyncpg = "^0.29.0"
psycopg2-binary = "^2.9.9"
pydantic = "^2.6.0"
pydantic-settings = "^2.1.0"
python-multipart = "^0.0.9"
pandas = "^2.2.0"
openpyxl = "^3.1.2"
qdrant-client = "^1.7.0"
sentence-transformers = "^2.3.1"
openai = "^1.12.0"
google-generativeai = "^0.7.0"
importlib-metadata = "^7.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
httpx = "^0.26.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
